pyproject.toml
[project]
name = "bot-telegram"
version = "0.1.0"
description = ""
authors = [
    {name = "Ivan",email = "ivanbozhkod@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "mindee (>=4.22.0,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)"
]

[tool.poetry]
packages = [{include = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
